requirements.txt
flask
google-cloud-storage
google-cloud-documentai
